ScopeNG: Check trace/trig mode currentIndex() for validity before casting
Multiple functions use a simple cast from the result of currentIndex()
to a projection type and assume everything will be fine. When running
sdrangel under ASAN, however, the following errors are logged which
indicate that the projection type is being initialized to an invalid
value at some point. The logged value corresponds to a currentIndex()
of -1, which is documented to be returned when the QComboBox either
has not been populated or has no current item set.

    ./sdrgui/gui/glscopenggui.cpp:1180:45: runtime error: load of value 4294967295, which is not a valid value for type 'ProjectionType'
    ./sdrgui/gui/glscopenggui.cpp:1189:19: runtime error: load of value 4294967295, which is not a valid value for type 'ProjectionType'
    ./sdrgui/dsp/scopevisng.h:45:12: runtime error: load of value 4294967295, which is not a valid value for type 'ProjectionType'
    ./sdrgui/dsp/scopevisng.h:753:73: runtime error: load of value 4294967295, which is not a valid value for type 'ProjectionType'
    ./sdrgui/dsp/scopevisng.h:780:55: runtime error: load of value 4294967295, which is not a valid value for type 'ProjectionType'
    ./sdrgui/dsp/scopevisng.h:45:12: runtime error: load of value 4294967295, which is not a valid value for type 'ProjectionType'
    ./sdrgui/dsp/scopevisng.cpp:807:21: runtime error: load of value 4294967295, which is not a valid value for type 'ProjectionType'
    ./sdrgui/dsp/scopevisng.h:97:12: runtime error: load of value 4294967295, which is not a valid value for type 'ProjectionType'
    ./sdrgui/dsp/scopevisng.h:97:12: runtime error: load of value 4294967295, which is not a valid value for type 'ProjectionType'
    ./sdrgui/dsp/scopevisng.h:551:66: runtime error: load of value 4294967295, which is not a valid value for type 'ProjectionType'
    ./sdrgui/dsp/scopevisng.h:553:47: runtime error: load of value 4294967295, which is not a valid value for type 'ProjectionType'
    ./sdrbase/dsp/projector.h:37:55: runtime error: load of value 4294967295, which is not a valid value for type 'ProjectionType'

It was assumed that the calls to fillTraceData() and fillTriggerData()
which occur prior to the combo boxes being filled were the cause, but
moving these function calls to occur afterwards results in segfaults
due to other non-initalized variables being accessed when the UI is
updated with the options. As an alternative, we replace the dumb casts
with a more intelligent helper function which will check the value for
validity and return a sane value if it is out of range.